some fixes

+++ zasulsky.egor/S1/Vector.hpp
@@ -26,7 +26,7 @@ namespace zasulsky
       data = new T[other.capacity];
       capacity = other.capacity;
       size = other.size;
-      for (int i = 0; i < size; ++i)
+      for (size_t i = 0; i < size; ++i)
       {
         data[i] = other.data[i];
       }
